WIP: Initial commit of order creation form
This is a super rough commit of the form being used to create a shopping
list. There are lots of things still not working/incomplete about it:
1. Prior to the table to add/remove items, the user should be presented
with an address page, but is not currently. This means the page with the
order won't work because there's no address in state
2. Authentication isn't in place yet so there's no user ID in state yet,
so API requests to /users/:id/lists from the app can't work yet. I
worked around this to make sure it worked by hard-coding a user and
address on line 3 of client/actions/lists.js
3. The styling is dogshit